Minior bug fix

+++ src/lean_dojo/interaction/dojo.py
@@ -425,7 +425,10 @@ class Dojo:
             repl_dst = Path(repl_file)
             with open("lakefile.lean", "a") as oup:
                 oup.write("\nlean_lib Lean4Repl {\n\n}\n")
-            os.remove("lakefile.olean")
+            if os.path.exists("lakefile.olean"):
+                os.remove("lakefile.olean")
+            if os.path.exists(".lake/lakefile.olean"):
+                os.remove(".lake/lakefile.olean")
 
         # Copy the REPL code to the right directory.
         repl_src = Path(__file__).with_name(repl_file)
